+++ core/htmlcov/d_48e9d7ebb482c3cc___init___py.html
@@ -60,12 +60,12 @@
             <button type="button" class="exc show_exc button_toggle_exc" value="exc" data-shortcut="x" title="Toggle lines excluded">0<span class="text"> excluded</span></button>
         </h2>
         <p class="text">
-            <a id="prevFileLink" class="nav" href="d_5cc1c4a2ff182d0d_apps_py.html">&#xab; prev</a> &nbsp; &nbsp;
+            <a id="prevFileLink" class="nav" href="d_48e9d7ebb482c3cc_0010_remove_review_dislikes_remove_review_likes_and_more_py.html">&#xab; prev</a> &nbsp; &nbsp;
             <a id="indexLink" class="nav" href="index.html">&Hat; index</a> &nbsp; &nbsp;
             <a id="nextFileLink" class="nav" href="d_5cc1c4a2ff182d0d_models_py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
-            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.1</a>,
-            created at 2022-06-13 16:27 +0200
+            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.2</a>,
+            created at 2022-08-02 19:06 +0200
         </p>
         <aside class="hidden">
             <button type="button" class="button_next_chunk" data-shortcut="j"/>
@@ -84,12 +84,12 @@
 <footer>
     <div class="content">
         <p>
-            <a id="prevFileLink" class="nav" href="d_5cc1c4a2ff182d0d_apps_py.html">&#xab; prev</a> &nbsp; &nbsp;
+            <a id="prevFileLink" class="nav" href="d_48e9d7ebb482c3cc_0010_remove_review_dislikes_remove_review_likes_and_more_py.html">&#xab; prev</a> &nbsp; &nbsp;
             <a id="indexLink" class="nav" href="index.html">&Hat; index</a> &nbsp; &nbsp;
             <a id="nextFileLink" class="nav" href="d_5cc1c4a2ff182d0d_models_py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
-            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.1</a>,
-            created at 2022-06-13 16:27 +0200
+            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.2</a>,
+            created at 2022-08-02 19:06 +0200
         </p>
     </div>
 </footer>
