+++ core/htmlcov/manage_py.html
@@ -62,10 +62,10 @@
         <p class="text">
             <a id="prevFileLink" class="nav" href="d_57760688d1f824db_urls_py.html">&#xab; prev</a> &nbsp; &nbsp;
             <a id="indexLink" class="nav" href="index.html">&Hat; index</a> &nbsp; &nbsp;
-            <a id="nextFileLink" class="nav" href="d_5cc1c4a2ff182d0d___init___py.html">&#xbb; next</a>
+            <a id="nextFileLink" class="nav" href="d_fe3e9668e34e13e2___init___py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
-            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.1</a>,
-            created at 2022-06-21 16:32 +0200
+            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.2</a>,
+            created at 2022-08-02 19:06 +0200
         </p>
         <aside class="hidden">
             <button type="button" class="button_next_chunk" data-shortcut="j"/>
@@ -107,10 +107,10 @@
         <p>
             <a id="prevFileLink" class="nav" href="d_57760688d1f824db_urls_py.html">&#xab; prev</a> &nbsp; &nbsp;
             <a id="indexLink" class="nav" href="index.html">&Hat; index</a> &nbsp; &nbsp;
-            <a id="nextFileLink" class="nav" href="d_5cc1c4a2ff182d0d___init___py.html">&#xbb; next</a>
+            <a id="nextFileLink" class="nav" href="d_fe3e9668e34e13e2___init___py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
-            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.1</a>,
-            created at 2022-06-21 16:32 +0200
+            <a class="nav" href="https://coverage.readthedocs.io">coverage.py v6.4.2</a>,
+            created at 2022-08-02 19:06 +0200
         </p>
     </div>
 </footer>
